requirements.txt
requests
beautifulsoup4
pytesseract
Pillow
scikit-learn
matplotlib
langchain 